chore: fix lint test build on main (#10866)
# Which Problems Are Solved

In order for golangci-lint to be able to compare what changed against
main, we have to check out main on branches.
If the pipeline runs on main, the branch can't be checked out twice and
the pipeline fails.

# How the Problems Are Solved

The step `Fetch main branch` in `lint_test_build` is only executed if
the workflow doesn't run on main, because in this case, the branch is
already checked out by the `Checkout Repository` step.

# Additional Changes

PRs can only be merged if `lint_test_buld` succeeds on an up-to-date
branch. When a PR is merged, it triggers a push event which runs
`lint_test_build` again on main with the same conditions. This is
obsolete. Pushes to main don't trigger the `lint_test_build` workflow
anymore.

# Additional Context

- Resolves
https://github.com/zitadel/zitadel/actions/runs/18339015014/job/52229685065

(cherry picked from commit 4e432a7e99bc32c385370b67a0334b8777a3b07d)